feat: LSP server
This adds a mix task to will start an LSP server for credo

```
mix credo.lsp
```

Utilizes the GenLSP library to factor out all of the LSP
specific bits.

This can be hooked up to Neovim 0.8+ using the following config

```lua
local credo = augroup("credo", { clear = true })

vim.api.nvim_create_autocmd({ "FileType" }, {
  group = credo,
  pattern = { "elixir" },
  callback = function()
    vim.lsp.start {
      name = "CredoLS",
      cmd = { "mix", "credo.lsp" },
      settings = {},
      root_dir = vim.fs.dirname(vim.fs.find({ "mix.exs", ".git" }, { upward = true })[1]),
    }
  end
})
```

Note: An optimization to be made for the Neovim config is to only start
it if the deteced mix.exs file has credo in it. This will attempt to
start it for any elixir project.